pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = "API Gateway Platform for Microservices"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
motor = "^3.3.2"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.26.0"
email-validator = "^2.1.0.post1"
python-multipart = "^0.0.7"
alembic = "^1.13.1"
loguru = "^0.7.2"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "4.0.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
black = "^23.12.0"
isort = "^5.13.1"
mypy = "^1.7.1"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
